# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * The social media platform
 */
export type CreateDownloadBodyPlatform =
  (typeof CreateDownloadBodyPlatform)[keyof typeof CreateDownloadBodyPlatform];

export const CreateDownloadBodyPlatform = {
  tiktok: "tiktok",
  instagram: "instagram",
  facebook: "facebook",
  youtube: "youtube",
  twitter: "twitter",
} as const;

export interface CreateDownloadBody {
  /** The social media video URL to download */
  url: string;
  /** The social media platform */
  platform: CreateDownloadBodyPlatform;
  /** User IP for tracking (optional) */
  userIp?: string;
}

export interface DownloadResult {
  success: boolean;
  /** Direct URL to download the video */
  downloadUrl: string;
  /** Suggested filename */
  filename: string;
  platform: string;
  /** Video quality (HD, etc.) */
  quality: string;
  /** Video title if available */
  title?: string;
  /** Thumbnail URL */
  thumbnail?: string;
  /** ID of the download record */
  downloadId?: number;
}

export type DownloadStatsByPlatform = { [key: string]: number };

export interface DownloadStats {
  totalDownloads: number;
  totalUsers: number;
  byPlatform: DownloadStatsByPlatform;
}

export interface AdminPinBody {
  pin: string;
}

export interface AdminVerifyResult {
  valid: boolean;
  message: string;
}

export type AdminDashboardByPlatform = { [key: string]: number };

export interface DownloadRecord {
  id: number;
  url: string;
  platform: string;
  userIp?: string;
  createdAt: string;
  filename?: string;
  success: boolean;
}

export interface DailyStat {
  date: string;
  count: number;
}

export interface AdminDashboard {
  totalDownloads: number;
  totalUsers: number;
  todayDownloads: number;
  byPlatform: AdminDashboardByPlatform;
  recentDownloads: DownloadRecord[];
  dailyStats: DailyStat[];
}

export interface AdminDownloadsList {
  downloads: DownloadRecord[];
  total: number;
  page: number;
  limit: number;
}

export interface ApiError {
  error: string;
  message: string;
}

export type GetAdminDashboardParams = {
  pin: string;
};

export type ListAdminDownloadsParams = {
  pin: string;
  page?: number;
  limit?: number;
};