# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Download a video from a social media URL
 * @summary Download a video
 */
export const CreateDownloadBody = zod.object({
  url: zod.string().describe("The social media video URL to download"),
  platform: zod
    .enum(["tiktok", "instagram", "facebook", "youtube", "twitter"])
    .describe("The social media platform"),
  userIp: zod.string().optional().describe("User IP for tracking (optional)"),
});

export const CreateDownloadResponse = zod.object({
  success: zod.boolean(),
  downloadUrl: zod.string().describe("Direct URL to download the video"),
  filename: zod.string().describe("Suggested filename"),
  platform: zod.string(),
  quality: zod.string().describe("Video quality (HD, etc.)"),
  title: zod.string().optional().describe("Video title if available"),
  thumbnail: zod.string().optional().describe("Thumbnail URL"),
  downloadId: zod.number().optional().describe("ID of the download record"),
});

/**
 * Get public download statistics
 * @summary Get download statistics
 */
export const GetDownloadStatsResponse = zod.object({
  totalDownloads: zod.number(),
  totalUsers: zod.number(),
  byPlatform: zod.record(zod.string(), zod.number()),
});

/**
 * Verify the owner panel PIN
 * @summary Verify admin PIN
 */
export const VerifyAdminPinBody = zod.object({
  pin: zod.string(),
});

export const VerifyAdminPinResponse = zod.object({
  valid: zod.boolean(),
  message: zod.string(),
});

/**
 * Get full dashboard data for the owner panel
 * @summary Get admin dashboard data
 */
export const GetAdminDashboardQueryParams = zod.object({
  pin: zod.coerce.string(),
});

export const GetAdminDashboardResponse = zod.object({
  totalDownloads: zod.number(),
  totalUsers: zod.number(),
  todayDownloads: zod.number(),
  byPlatform: zod.record(zod.string(), zod.number()),
  recentDownloads: zod.array(
    zod.object({
      id: zod.number(),
      url: zod.string(),
      platform: zod.string(),
      userIp: zod.string().optional(),
      createdAt: zod.string(),
      filename: zod.string().optional(),
      success: zod.boolean(),
    }),
  ),
  dailyStats: zod.array(
    zod.object({
      date: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * List all download records for admin view
 * @summary List all downloads
 */
export const listAdminDownloadsQueryPageDefault = 1;
export const listAdminDownloadsQueryLimitDefault = 20;

export const ListAdminDownloadsQueryParams = zod.object({
  pin: zod.coerce.string(),
  page: zod.coerce.number().default(listAdminDownloadsQueryPageDefault),
  limit: zod.coerce.number().default(listAdminDownloadsQueryLimitDefault),
});

export const ListAdminDownloadsResponse = zod.object({
  downloads: zod.array(
    zod.object({
      id: zod.number(),
      url: zod.string(),
      platform: zod.string(),
      userIp: zod.string().optional(),
      createdAt: zod.string(),
      filename: zod.string().optional(),
      success: zod.boolean(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  limit: zod.number(),
});