# lib/api-zod/src/generated/types/adminDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminDashboardByPlatform } from "./adminDashboardByPlatform";
import type { DailyStat } from "./dailyStat";
import type { DownloadRecord } from "./downloadRecord";

export interface AdminDashboard {
  totalDownloads: number;
  totalUsers: number;
  todayDownloads: number;
  byPlatform: AdminDashboardByPlatform;
  recentDownloads: DownloadRecord[];
  dailyStats: DailyStat[];
}