# lib/api-zod/src/generated/types/adminDownloadsList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadRecord } from "./downloadRecord";

export interface AdminDownloadsList {
  downloads: DownloadRecord[];
  total: number;
  page: number;
  limit: number;
}