# lib/api-zod/src/generated/types/adminVerifyResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminVerifyResult {
  valid: boolean;
  message: string;
}