# lib/api-zod/src/generated/types/createDownloadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateDownloadBodyPlatform } from "./createDownloadBodyPlatform";

export interface CreateDownloadBody {
  /** The social media video URL to download */
  url: string;
  /** The social media platform */
  platform: CreateDownloadBodyPlatform;
  /** User IP for tracking (optional) */
  userIp?: string;
}