# lib/api-zod/src/generated/types/dailyStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyStat {
  date: string;
  count: number;
}