# lib/api-zod/src/generated/types/downloadRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadRecord {
  id: number;
  url: string;
  platform: string;
  userIp?: string;
  createdAt: string;
  filename?: string;
  success: boolean;
}