# lib/api-zod/src/generated/types/downloadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadResult {
  success: boolean;
  /** Direct URL to download the video */
  downloadUrl: string;
  /** Suggested filename */
  filename: string;
  platform: string;
  /** Video quality (HD, etc.) */
  quality: string;
  /** Video title if available */
  title?: string;
  /** Thumbnail URL */
  thumbnail?: string;
  /** ID of the download record */
  downloadId?: number;
}