# lib/api-zod/src/generated/types/downloadStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * URL Downloader API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DownloadStatsByPlatform } from "./downloadStatsByPlatform";

export interface DownloadStats {
  totalDownloads: number;
  totalUsers: number;
  byPlatform: DownloadStatsByPlatform;
}